+++ lib/src/ledger/ledger_gatt_reader.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'dart:typed_data';
 
-import 'package:ledger_flutter/src/exceptions/ledger_exception.dart';
-import 'package:ledger_flutter/src/utils/buffer.dart';
+import 'package:ledger/src/exceptions/ledger_exception.dart';
+import 'package:ledger/src/utils/buffer.dart';
 
 class LedgerGattReader {
   /// The APDU command tag 0x05 is used to transfer application specific data.