+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* Support more Ledger devices
+* Update example to use Ethereum Ledger App
+* Update package name
+
 ## 1.0.1
 
 * Changed package name